fix: Address CodeRabbit review issues and improve type safety
- Remove redundant unique index on food_id (already has unique constraint)
- Fix meals.kcal type mismatch: change migration from Integer to Float
- Add return type hints to all test functions in test_app_coverage_gaps.py
- Rename unused test_environment fixture to _test_environment
- Add return type hints to mock_glob functions in test_check_failing_tests_unit.py
- Move sqlalchemy.exc import to top in test_core_db_additional_coverage.py
- Use cast() instead of type: ignore for wrapper functions in routers

Fixes all CodeRabbit review comments and improves type safety

+++ alembic/versions/202501110001_add_nutrition_tables.py
@@ -74,7 +74,9 @@ def upgrade() -> None:
         sa.Column("recipe_id", sa.Integer(), nullable=True),
         sa.Column("title", sa.String(length=500), nullable=False),
         sa.Column("title_translated", sa.String(length=500), nullable=True),
-        sa.Column("kcal", sa.Integer(), nullable=False),
+        sa.Column(
+            "kcal", sa.Float(), nullable=False
+        ),  # Changed from Integer to Float to match model
         sa.Column("protein_g", sa.Float(), nullable=False, server_default="0.0"),
         sa.Column("fat_g", sa.Float(), nullable=False, server_default="0.0"),
         sa.Column("carbs_g", sa.Float(), nullable=False, server_default="0.0"),
